requirements.txt
torch==2.1.0
torchvision==0.15.0
matplotlib==3.7.1
numpy==1.23.5
scikit-learn==1.3.2
seaborn==0.12.2
onnx==1.15.0
onnxruntime==1.16.0
tqdm==4.66.0
opencv-python==4.7.0.72
streamlit==1.24.0
streamlit-drawable-canvas==0.10.0
Pillow==9.4.0
